Remove DBALException
